Moved source for replay to replay sub-directory of tools
