+++ tools/replay/ReplayOperationManager.inl
